abstraction in ConfigSerializer
