+++ emotesServer/src/main/java/io/github/kosmx/emotes/server/config/ConfigSerializer.java
@@ -1,9 +1,8 @@
-package io.github.kosmx.emotes.main.config;
+package io.github.kosmx.emotes.server.config;
 
 import com.google.gson.*;
-import io.github.kosmx.emotes.common.tools.Pair;
-import io.github.kosmx.emotes.main.EmoteHolder;
 import io.github.kosmx.emotes.common.SerializableConfig;
+import io.github.kosmx.emotes.common.tools.Pair;
 import io.github.kosmx.emotes.executor.EmoteInstance;
 
 import java.lang.reflect.Type;
@@ -14,7 +13,7 @@ public class ConfigSerializer implements JsonDeserializer<SerializableConfig>, J
     @Override
     public SerializableConfig deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException{
         JsonObject node = json.getAsJsonObject();
-        SerializableConfig config = EmoteInstance.instance.isClient() ? new ClientConfig() : new SerializableConfig();
+        SerializableConfig config = this.newConfig();
         config.configVersion = SerializableConfig.staticConfigVersion;
         if(node.has("config_version"))config.configVersion = node.get("config_version").getAsInt();
         if(config.configVersion < SerializableConfig.staticConfigVersion){
@@ -26,8 +25,11 @@ public class ConfigSerializer implements JsonDeserializer<SerializableConfig>, J
 
         config.iterate(entry -> deserializeEntry(entry, node));
 
-        if(EmoteInstance.instance.isClient()) clientDeserialize(node, (ClientConfig) config);
         return config;
+    }
+
+    protected SerializableConfig newConfig(){
+        return new SerializableConfig();
     }
 
     private void deserializeEntry(SerializableConfig.ConfigEntry<?> entry, JsonObject node){
@@ -49,34 +51,11 @@ public class ConfigSerializer implements JsonDeserializer<SerializableConfig>, J
         }
     }
 
-    private void clientDeserialize(JsonObject node, ClientConfig config){
-        EmoteFixer emoteFixer = new EmoteFixer(config.configVersion);
-        if(node.has("fastmenu")) fastMenuDeserializer(node.get("fastmenu").getAsJsonObject(), config, emoteFixer);
-        if(node.has("keys")) keyBindsDeserializer(node.get("keys").getAsJsonArray(), config, emoteFixer);
-    }
-
-    private void fastMenuDeserializer(JsonObject node, ClientConfig config, EmoteFixer fixer){
-        for(int i = 0; i != 8; i++){
-            if(node.has(Integer.toString(i))){
-                config.fastMenuHash[i] = fixer.getEmoteID(node.get(Integer.toString(i)));
-            }
-        }
-    }
-
-    private void keyBindsDeserializer(JsonArray node, ClientConfig config, EmoteFixer fixer){
-        for(JsonElement object : node){
-            JsonObject n = object.getAsJsonObject();
-            config.emotesWithHash.add(new Pair<>(fixer.getEmoteID(n.get("id")), n.get("key").getAsString()));
-            //keyBindDeserializer(object.getAsJsonObject());
-        }
-    }
-
     @Override
     public JsonElement serialize(SerializableConfig config, Type typeOfSrc, JsonSerializationContext context){
         JsonObject node = new JsonObject();
         node.addProperty("config_version", SerializableConfig.staticConfigVersion); //I always save config with the latest format.
         config.iterate(entry -> serializeEntry(entry, node));
-        if(config instanceof ClientConfig) clientSerialize((ClientConfig) config, node);
         return node;
     }
 
@@ -87,36 +66,6 @@ public class ConfigSerializer implements JsonDeserializer<SerializableConfig>, J
         else if(entry instanceof SerializableConfig.FloatConfigEntry){
             node.addProperty(entry.getName(), (float)((SerializableConfig.FloatConfigEntry) entry).get());
         }
-    }
-
-    private void clientSerialize(ClientConfig config, JsonObject node){
-        node.add("fastmenu", fastMenuSerializer(config));
-        node.add("keys", keyBindsSerializer(config));
-    }
-
-    private JsonObject fastMenuSerializer(ClientConfig config){
-        JsonObject node = new JsonObject();
-        for(int i = 0; i != 8; i++){
-            if(config.fastMenuEmotes[i] != null){
-                node.addProperty(Integer.toString(i), config.fastMenuEmotes[i].hash);
-            }
-        }
-        return node;
-    }
-
-    private JsonArray keyBindsSerializer(ClientConfig config){
-        JsonArray array = new JsonArray();
-        for(EmoteHolder emote : config.emotesWithKey){
-            array.add(keyBindSerializer(emote));
-        }
-        return array;
-    }
-
-    private JsonObject keyBindSerializer(EmoteHolder emote){
-        JsonObject node = new JsonObject();
-        node.addProperty("id", emote.hash);
-        node.addProperty("key", emote.keyBinding.getTranslationKey());
-        return node;
     }
 
 }
